Changes upper bounds of logkappa2
using VAST_3.9.0 the upper bound of logkappa2 is reached
using the default settings b/c of the small number of mesh points.
This commit changes the upper bound to 10 using
run_model = FALSE to get the correct map.
This trick can be found in
https://github.com/James-Thorson-NOAA/VAST/issues/300
which was started by @kellijohnson-NOAA b/c of similar issues
when running the model for West Coast species.