fix(configuration): Allow for Tweedie by
changing spatiotemporal1 and spatiotemporal2 to a single column
with a quoted R command that is evaluated. I am happy to take changes
to this solution if someone knows how to do psuedo evaluation like how
tidyverse does it with equos and quos and all that stuff but for right
now this is working for me.

Part of #30