fix(filter_boundaries): Only filter what is present
Thanks to @chantelwetzel-noaa for letting me know that there was a bug.
The filtering was too complicated and it would look for boundaries that
were not present. Additionally, there was an error if a data point was
directly on a boundary line leading to the northern and southern lines
being equal. New checks are in place and the code is simpler.

Close #33